src/files/requirements.txt
aiofiles==24.1.0
alembic==1.13.2
asyncpg==0.29.0
cashews==7.3.1
environs==11.0.0
grpcio==1.66.1
grpc-accesslog==0.3.0
grpcio_tools==1.66.1
redis==5.0.8
sqlalchemy==2.0.34
uvloop==0.21.0
